.h file changes are done

+++ 3_Implementation/inc/calculator_operations.h
@@ -16,7 +16,7 @@
 * @param[in] operand2 
 * @return Result of operand1 + operand2
 */
-int add(int operand1, int operand2);
+float add(float operand1, float operand2);
 
 /**
 *  subtracts the operand1, operand2 and returns the result
@@ -24,7 +24,7 @@ int add(int operand1, int operand2);
 * @param[in] operand2 
 * @return Result of operand1 - operand2
 */
-int subtract(int operand1, int operand2);
+float subtract(float operand1, float operand2);
 
 /**
 *  multiply the operand1, operand2 and returns the result
@@ -32,7 +32,7 @@ int subtract(int operand1, int operand2);
 * @param[in] operand2 
 * @return Result of operand1 * operand2
 */
-int multiply(int operand1, int operand2);
+float multiply(float operand1, float operand2);
 
 
 /**
@@ -42,6 +42,6 @@ int multiply(int operand1, int operand2);
 * @return integer value of the operand1 / operand2
 * @note returns 0 for divide by 0 error
 */
-int divide(int operand1, int operand2);
+float divide(float operand1, float operand2);
 
 #endif  /* #define __CALCULATOR_OPERATIONS_H__ */
